third_party/blink/web_tests/html/dialog/form-method-dialog-expected.txt

CONSOLE WARNING: line 177: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests form submission with method=dialog

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Clicking outer-submit-empty-string
PASS outer; dialog.open is false
PASS outer; dialog.returnValue is ""
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking outer-submit-no-value
PASS outer; dialog.open is false
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking inner-submit-yes
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is "Yes"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking inner-submit-no
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is "No"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking no-dialog-ancestor-1
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking no-dialog-ancestor-2
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking host-submit-yes
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is true
PASS inner; dialog.returnValue is "init"
PASS host; dialog.open is false
PASS host; dialog.returnValue is "Yes"


Submitting a form without submit button (this should not crash)


Clicking a button in a closed dialog
PASS dialog.open is false
PASS dialog.returnValue is "init"


Activating an image button by click()
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is "0,0"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Activating an image button by keyboard
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is "0,0"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"


Clicking an image button
PASS outer; dialog.open is true
PASS outer; dialog.returnValue is "init"
PASS inner; dialog.open is false
PASS inner; dialog.returnValue is "10,5"
PASS host; dialog.open is true
PASS host; dialog.returnValue is "init"
PASS successfullyParsed is true

TEST COMPLETE

 
